requirements.txt
streamlit
tensorflow
opencv-python-headless
Pillow
scikit-learn